03-embedding-ms/requirements.txt
# Common dependencies are in shared/requirements.txt
google-genai
flask
numpy